+++ lib/oAuth/client.js
@@ -1,9 +1,5 @@
-const randomString = require('randomstring')
 const simpleOauth2 = require('simple-oauth2')
-const {
-  CMS_MESSAGE_ERROR,
-  CMS_MESSAGE_SUCCESS
-} = require('./constants')
+const { CMS_MESSAGE_ERROR, CMS_MESSAGE_SUCCESS } = require('./config')
 
 class OAuth2Client {
   static constructResponse (message, content) {
@@ -12,8 +8,6 @@ class OAuth2Client {
 
   constructor (provider, credentials, redirectUri = null, scope = null) {
     this.instance = simpleOauth2.create(credentials)
-
-    this.state = null
 
     this.config = {
       provider,
@@ -32,35 +26,18 @@ class OAuth2Client {
 
   constructErrorResponse (error) {
     const getResponse = err => OAuth2Client.constructResponse(CMS_MESSAGE_ERROR, err)
-    if (error && error.message) getResponse(error)
+    if (error && error.message) return getResponse(error)
     return getResponse({ message: error })
   }
 
-  getProvider () {
-    return this.config.provider
-  }
+  getProvider () { return this.config.provider }
 
-  getStatesEqual (state) {
-    return state === this.state
-  }
-
-  /**
-   * Return authorizeUrl to initiate oAuth-flow
-   * @return {String} url that will be redirected to
-   */
-  authorizeUrl () {
+  authorizeUrl (uuid) {
     const { redirectUri: redirect_uri, scope } = this.config
-    const state = randomString.generate(32)
-    this.state = state
-    return this.instance.authorizationCode.authorizeURL({ redirect_uri, scope, state })
+    return this.instance.authorizationCode.authorizeURL({ redirect_uri, scope, state: uuid })
   }
 
-  /**
-   * Retrieves access_token from oAuth response, handles errors returned from the API
-   * @param  {Object} query response query
-   * @return {Promise<Object>} resolves to netlify-cms-compatible response object
-   */
-  async accessToken (query) {
+  async accessToken (uuid, query) {
     const { redirectUri: redirect_uri } = this.config
 
     const {
@@ -70,7 +47,7 @@ class OAuth2Client {
       state
     } = query
 
-    if (!this.getStatesEqual(state)) return this.constructErrorResponse('State does not match supplied value')
+    if (state !== uuid) return this.constructErrorResponse('CSRF detected')
     if (error && errorDescription) return this.constructErrorResponse(errorDescription)
 
     try {
