+++ lib/oAuth/config.js
@@ -1,4 +1,3 @@
-
 const CMS_MESSAGE_SUCCESS = 'success'
 const CMS_MESSAGE_ERROR = 'error'
