+++ lib/oAuth/factory.js
@@ -1,7 +1,7 @@
 require('dotenv').config({ silent: true })
 
-const OAuth2Client = require('./oAuth2Client')
-const { BITBUCKET_SETTINGS, GITLAB_SETTINGS, GITHUB_SETTINGS } = require('./constants')
+const OAuth2Client = require('./client')
+const { BITBUCKET_SETTINGS, GITLAB_SETTINGS, GITHUB_SETTINGS } = require('./config')
 
 const DEFAULT_PROVIDER = 'github'
